# swagger/v1/schema.yaml
...
openapi: 3.1.0
info:
  title: Reckoning API
  version: v1
  license:
    name: GNU General Public License v3.0
    url: https://github.com/mortik/reckoning/blob/main/LICENSE
paths:
  "/account":
    get:
      summary: The signed-in user's account
      tags:
      - Account
      operationId: account
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Account"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
    patch:
      summary: Update account settings
      tags:
      - Account
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AccountInput"
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Account"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/customers":
    get:
      summary: Customers list
      tags:
      - Customers
      operationId: customers
      parameters:
      - "$ref": "#/components/parameters/PageParameter"
      - "$ref": "#/components/parameters/PerPageParameter"
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Customers"
          headers:
            Link:
              description: 'RFC 8288 pagination links: self, first, prev, next, last.'
              schema:
                type: string
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
    post:
      summary: Create new Customer
      tags:
      - Customers
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CustomerInput"
      responses:
        '201':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Customer"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
  "/customers/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      required: true
    get:
      summary: Get Customer
      tags:
      - Customers
      operationId: customer
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Customer"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          "$ref": "#/components/responses/SchemaValidationError"
    patch:
      summary: Update Customer
      tags:
      - Customers
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CustomerInput"
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Customer"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
    delete:
      summary: Destroy Customer
      tags:
      - Customers
      operationId: destroyCustomer
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/expenses":
    get:
      summary: Expenses list
      tags:
      - Expenses
      operationId: expenses
      parameters:
      - "$ref": "#/components/parameters/PageParameter"
      - "$ref": "#/components/parameters/PerPageParameter"
      - name: year
        in: query
        required: false
        schema:
          type: integer
      - name: quarter
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 4
      - name: month
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 12
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: query
        in: query
        required: false
        description: Free-text search over description and seller.
        schema:
          type: string
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Expenses"
          headers:
            Link:
              description: RFC 8288 pagination links.
              schema:
                type: string
        '403':
          description: expenses not enabled for this account
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          "$ref": "#/components/responses/SchemaValidationError"
    post:
      summary: Create new Expense
      tags:
      - Expenses
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ExpenseInput"
      responses:
        '201':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Expense"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '403':
          description: expenses not enabled for this account
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/expenses/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      required: true
    get:
      summary: Get Expense
      tags:
      - Expenses
      operationId: expense
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Expense"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          "$ref": "#/components/responses/SchemaValidationError"
    patch:
      summary: Update Expense
      tags:
      - Expenses
      operationId: updateExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ExpenseInput"
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Expense"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
    delete:
      summary: Destroy Expense
      tags:
      - Expenses
      operationId: destroyExpense
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Message"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
  "/invoices":
    get:
      summary: Invoices list
      tags:
      - Invoices
      operationId: invoices
      parameters:
      - "$ref": "#/components/parameters/PageParameter"
      - "$ref": "#/components/parameters/PerPageParameter"
      - name: state
        in: query
        required: false
        schema:
          type: string
          enum:
          - created
          - charged
          - paid
      - name: year
        in: query
        required: false
        schema:
          type: integer
      - name: quarter
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 4
      - name: month
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 12
      - name: paid_in_year
        in: query
        required: false
        schema:
          type: integer
      - name: paid_in_quarter
        in: query
        required: false
        schema:
          type: integer
      - name: paid_in_month
        in: query
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Invoices"
          headers:
            Link:
              description: RFC 8288 pagination links.
              schema:
                type: string
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '400':
          "$ref": "#/components/responses/SchemaValidationError"
    post:
      summary: Create new Invoice
      tags:
      - Invoices
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/InvoiceInput"
      responses:
        '201':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Invoice"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ValidationError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
servers:
- url: "/api/v1"
  description: Current host
components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type:
          - string
          - 'null'
        subdomain:
          type:
          - string
          - 'null'
        plan:
          type:
          - string
          - 'null'
        vatId:
          type:
          - string
          - 'null'
        tax:
          type:
          - string
          - 'null'
          description: Default tax rate, stored as a string.
        provision:
          type:
          - string
          - 'null'
        bank:
          type:
          - string
          - 'null'
        accountNumber:
          type:
          - string
          - 'null'
        bankCode:
          type:
          - string
          - 'null'
        iban:
          type:
          - string
          - 'null'
        bic:
          type:
          - string
          - 'null'
        defaultFrom:
          type:
          - string
          - 'null'
        signature:
          type:
          - string
          - 'null'
        address:
          type:
          - string
          - 'null'
        country:
          type:
          - string
          - 'null'
        publicEmail:
          type:
          - string
          - 'null'
        telefon:
          type:
          - string
          - 'null'
        fax:
          type:
          - string
          - 'null'
        website:
          type:
          - string
          - 'null'
        officeSpace:
          type:
          - integer
          - 'null'
        deductibleOfficeSpace:
          type:
          - integer
          - 'null'
        offerHeadline:
          type:
          - string
          - 'null'
        featureExpenses:
          type:
          - boolean
          - 'null'
        featureLogbook:
          type:
          - boolean
          - 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - createdAt
      - updatedAt
      title: Account
    AccountInput:
      type: object
      properties:
        name:
          type:
          - string
          - 'null'
        subdomain:
          type:
          - string
          - 'null'
        vatId:
          type:
          - string
          - 'null'
        tax:
          type:
          - string
          - 'null'
        provision:
          type:
          - string
          - 'null'
        bank:
          type:
          - string
          - 'null'
        accountNumber:
          type:
          - string
          - 'null'
        bankCode:
          type:
          - string
          - 'null'
        iban:
          type:
          - string
          - 'null'
        bic:
          type:
          - string
          - 'null'
        defaultFrom:
          type:
          - string
          - 'null'
        signature:
          type:
          - string
          - 'null'
        address:
          type:
          - string
          - 'null'
        country:
          type:
          - string
          - 'null'
        publicEmail:
          type:
          - string
          - 'null'
        telefon:
          type:
          - string
          - 'null'
        fax:
          type:
          - string
          - 'null'
        website:
          type:
          - string
          - 'null'
        officeSpace:
          type:
          - integer
          - 'null'
        deductibleOfficeSpace:
          type:
          - integer
          - 'null'
        offerHeadline:
          type:
          - string
          - 'null'
      additionalProperties: false
      title: AccountInput
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type:
          - string
          - 'null'
        country:
          type:
          - string
          - 'null'
        email:
          type:
          - string
          - 'null'
        telefon:
          type:
          - string
          - 'null'
        fax:
          type:
          - string
          - 'null'
        website:
          type:
          - string
          - 'null'
        paymentDue:
          type:
          - integer
          - 'null'
        invoiceEmail:
          type:
          - string
          - 'null'
        invoiceEmailCc:
          type:
          - string
          - 'null'
        invoiceEmailBcc:
          type:
          - string
          - 'null'
        defaultFrom:
          type:
          - string
          - 'null'
        emailTemplate:
          type:
          - string
          - 'null'
        offerDisclaimer:
          type:
          - string
          - 'null'
        employmentDate:
          type:
          - string
          - 'null'
          format: date
        employmentEndDate:
          type:
          - string
          - 'null'
          format: date
        weeklyHours:
          type:
          - integer
          - 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - name
      - createdAt
      - updatedAt
      title: Customer
    CustomerInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        country:
          type: string
        email:
          type: string
        telefon:
          type: string
        fax:
          type: string
        website:
          type: string
        paymentDue:
          type: integer
        invoiceEmail:
          type: string
        invoiceEmailCc:
          type: string
        invoiceEmailBcc:
          type: string
        defaultFrom:
          type: string
        emailTemplate:
          type: string
        offerDisclaimer:
          type: string
        employmentDate:
          type: string
          format: date
        employmentEndDate:
          type: string
          format: date
        weeklyHours:
          type: integer
      additionalProperties: false
      required:
      - name
      title: CustomerInput
    Customers:
      type: array
      items:
        "$ref": "#/components/schemas/Customer"
      title: Customers
...
    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        expenseType:
          type: string
        description:
          type:
          - string
          - 'null'
        seller:
          type:
          - string
          - 'null'
        date:
          type:
          - string
          - 'null'
          format: date
        value:
          type:
          - string
          - 'null'
        vatPercent:
          type:
          - integer
          - 'null'
        vatValue:
          type:
          - string
          - 'null'
        privateUsePercent:
          type:
          - integer
          - 'null'
        interval:
          type:
          - string
          - 'null'
          description: once, monthly, quarterly, yearly.
        startedAt:
          type:
          - string
          - 'null'
          format: date
        endedAt:
          type:
          - string
          - 'null'
          format: date
        afaTypeId:
          type:
          - string
          - 'null'
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - expenseType
      - createdAt
      - updatedAt
      title: Expense
...
    ExpenseInput:
      type: object
      properties:
        expense_type:
          type: string
        afa_type_id:
          type:
          - string
          - 'null'
          format: uuid
        description:
          type:
          - string
          - 'null'
        seller:
          type:
          - string
          - 'null'
        date:
          type:
          - string
          - 'null'
          format: date
        value:
          type:
          - string
          - number
          - 'null'
        private_use_percent:
          type:
          - integer
          - 'null'
        vat_percent:
          type:
          - integer
          - 'null'
        interval:
          type:
          - string
          - 'null'
        started_at:
          type:
          - string
          - 'null'
          format: date
        ended_at:
          type:
          - string
          - 'null'
          format: date
      additionalProperties: false
      title: ExpenseInput
    Expenses:
      type: array
      items:
        "$ref": "#/components/schemas/Expense"
      title: Expenses
...
    InvoiceInput:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        ref:
          type:
          - integer
          - 'null'
        date:
          type:
          - string
          - 'null'
          format: date
        delivery_date:
          type:
          - string
          - 'null'
          format: date
        payment_due_date:
          type:
          - string
          - 'null'
          format: date
        positions_attributes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              description:
                type:
                - string
                - 'null'
              hours:
                type:
                - string
                - number
                - 'null'
              rate:
                type:
                - string
                - number
                - 'null'
              value:
                type:
                - string
                - number
                - 'null'
              timer_ids:
                type: array
                items:
                  type: string
                  format: uuid
              _destroy:
                type: boolean
            additionalProperties: false
      additionalProperties: false
      title: InvoiceInput
...
    Invoices:
      type: array
      items:
        "$ref": "#/components/schemas/Invoice"
      title: Invoices
...
    Message:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
      required:
      - message
      title: Message
...
    StandardError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
      required:
      - message
      title: StandardError
...
    ValidationError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      additionalProperties: false
      required:
      - code
      - message
      title: ValidationError
  parameters:
    PageParameter:
      name: page
      in: query
      required: false
      description: 1-based page number. Page boundaries are described by the Link
        response header.
      schema:
        type: integer
        minimum: 1
    PerPageParameter:
      name: perPage
      in: query
      required: false
      description: Page size, or "all" to skip pagination entirely. Values above the
        endpoint's maximum are rejected with 400.
      schema:
        oneOf:
        - type: integer
          minimum: 1
        - type: string
          enum:
          - all
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /sessions. Used by non-browser clients.
    SessionCookie:
      type: apiKey
      in: cookie
      name: RECKONING_TEST
      description: Devise session cookie. Used by the web frontend.
  responses:
    SchemaValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
            required:
            - error
            - details
security: []